basic integration implemented
connection between flask, sqlite, and react have been implemented.
tailwind is the frontend of choice.

[ACCOUNT]
currently, only the ability to create an account has been implemented, with no visual feedback. there is a reply that can be read depicting outcome of this operation in the form of a json response, but has not been implemented.

logging in and login persistence has not been implemented.